requirements.txt
python-dotenv
psycopg2-binary
SQLAlchemy
pandas
treelib
colorama
openai
mysql-connector-python